+++ SendBirdSDK.framework/Headers/SBDAppInfo.h
@@ -41,6 +41,12 @@
 @property (assign, atomic, readonly) BOOL useReaction;
 
 /**
+ * This is the application attributes list using on your Application ID.
+ * @since 3.0.198
+ */
+@property (strong, nonatomic, readonly, nullable) NSArray<NSString *> *applicationAttributes;
+
+/**
  * This function can check if Emoji information needs to be updated to date.
  *
  * @param prevEmojiHash Emoji hash value in use